VLMs/requirements_1.txt
auto_mix_prep==0.2.0
braceexpand==0.1.7
clip==1.0
fairscale==0.4.4
fvcore==0.1.5.post20221221
matplotlib==3.7.1
numpy==1.23.5
open_clip_torch==2.23.0
opencv_python==4.8.1.78
opencv_python_headless==4.7.0.72
packaging==24.1
pandas==2.2.2
Pillow==10.3.0
pycocotools==2.0.7
Requests==2.32.3
ruamel.base==1.0.0
scipy==1.13.1
seaborn==0.13.2
timm==0.6.13
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.37.2
VL_CheckList==0.0.1
vsr==1.0.6.1
webdataset==0.2.57
